+++ packages/client/src/form/onSubmit.ts
@@ -1,11 +1,19 @@
 import { useRouter } from "@tanstack/react-router";
-import type { ShapeSchema } from "@use-pico/common";
+import {
+	cleanOf,
+	ErrorSchema,
+	mapEmptyToNull,
+	onAxiosSchemaError,
+	type ShapeSchema,
+	withErrors,
+} from "@use-pico/common";
 import type { z } from "zod";
 import type { Form } from "./Form";
 
 export namespace onSubmit {
 	export interface Props<TShapeSchema extends ShapeSchema> {
 		mutation: Form.Props.Mutation<TShapeSchema>;
+		onError?(error: string): void;
 		/**
 		 * Map form values to mutation request values (output of this goes directly into mutation).
 		 *
@@ -30,55 +38,47 @@ export namespace onSubmit {
 
 export const onSubmit = <TShapeSchema extends ShapeSchema>({
 	mutation,
+	onError,
 	map = ({ cleanup }) => {
 		return cleanup();
 	},
 }: onSubmit.Props<TShapeSchema>) => {
 	const router = useRouter();
-	// const submit = form.handleSubmit(async (values) => {
-	// 	return mutation
-	// 		.mutateAsync(
-	// 			await map({
-	// 				values,
-	// 				cleanup() {
-	// 					return cleanOf(mapEmptyToNull(values));
-	// 				},
-	// 			}),
-	// 			{
-	// 				onSuccess() {
-	// 					router.invalidate();
-	// 				},
-	// 				onError(error) {
-	// 					withErrors({
-	// 						error,
-	// 						errors: [
-	// 							onAxiosSchemaError({
-	// 								error,
-	// 								schema: ErrorSchema,
-	// 								onError: ({ data }) => {
-	// 									form.setError("root", {
-	// 										message: data.message,
-	// 									});
-	// 								},
-	// 							}),
-	// 						],
-	// 						onError(error) {
-	// 							console.log("Error", error);
-	// 							form.setError("root", {
-	// 								message: error.message,
-	// 							});
-	// 						},
-	// 					});
-	// 				},
-	// 			},
-	// 		)
-	// 		.catch(() => {
-	// 			//
-	// 		});
-	// });
-
-	return (e: any) => {
-		e.stopPropagation();
-		// submit(e);
+	return async (values: z.infer<TShapeSchema>) => {
+		return mutation
+			.mutateAsync(
+				await map({
+					values,
+					cleanup() {
+						return cleanOf(mapEmptyToNull(values));
+					},
+				}),
+				{
+					onSuccess() {
+						router.invalidate();
+					},
+					onError(error) {
+						withErrors({
+							error,
+							errors: [
+								onAxiosSchemaError({
+									error,
+									schema: ErrorSchema,
+									onError: ({ data }) => {
+										onError?.(data.message);
+									},
+								}),
+							],
+							onError(error) {
+								console.log("Error", error);
+								onError?.(error.message);
+							},
+						});
+					},
+				},
+			)
+			.catch(() => {
+				//
+			});
 	};
 };